Use reflection to parse values (#78)
* stated on generic struct mapper

* added todos

* added better tests

* tested creating struct

* added struct tags

* fixing tests

* fixing final tests

* added example of Struct

* added more convenience to create structs from account

* move from camelCase to snake_case

* remove println

* remove the type since it can be derrived from value

* fixed nested struct support

* fixed NFT contract and added MetadataViews contract

* added metadata structs except royalties with caps and some shared ones

* remove WithStructArg

* added option to skip field in struct if we do not want it in cadence

* use json structtag if it is there and cadence is not there

* tidy

+++ transaction_v3_test.go
@@ -58,7 +58,7 @@ transaction(test:UInt64) {
     log(test)
  }
 }
-`, WithArg("test", 1), WithSignerServiceAccount())
+`, WithArg("test", uint64(1)), WithSignerServiceAccount())
 		assert.NoError(t, res.Err)
 	})
